requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.6.2
sqlmodel==0.0.8
alembic==1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sentence-transformers==2.2.2
chromadb==0.4.3
rapidfuzz==3.16.1
httpx==0.26.3
python-multipart==0.0.6
redis==5.3.2
celery==5.3.1
